Reorganize testing code
