+++ tests/test_sampling/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Tests for sampling."""
